refactored tools folder structure
