+++ cmd/imagine/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/pilosa/tools/imagine"
+)
+
+func main() {
+	imagine.NewConfig().Execute()
+}
