+++ dx/solo_query.go
