+++ imagine/enums_cachetype.go
