+++ imagine/enums_densitytype.go
