+++ imagine/enums_fieldtype.go
