+++ imagine/enums_stamptype.go
