+++ imagine/enums_timequantum.go
