+++ imagine/enums_verifytype.go
