+++ imagine/generators_test.go
