+++ imagine/spec.go
